BeadData struct now knows its beads residue_type
To allow for this, the enum and vector of strings now have G as the last
type. G should never be assigned as the bead_type, but can be assigned
to the new field residue_type. residue_type joins the location field in
having a default value in the constructor, allowing for both fields to
be ignored in getBeadID lookups.